Cambio de formato
Cambio de formato de registro de Form a Dialog
